requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
sqlalchemy_utils
alembic
psycopg2-binary